Last use of Paris Observatory nutation factors
I have found what appears to be an error in the sign of what Meeus calls
the M coefficient of the term whose Delta psi coefficient is -16. I have
used the source of the IAU's nut80c.c file as a tie-breaker in this, and
it agrees with Meeus. Or perhaps I should say Meeus agrees with it. The
file comes from http://www.iausofa.org/2018_0130_C/sofa_c-20180130.tar.gz