Tweak planetary position code.
At the moment it looks to me like the difference between my results and
Meeus' worked example for the position of Venus is in the actual
coordinate transform from ecliptic to equatorial. Which is odd because
it is just a couple atan2()s and a sqrt(). I did ditch a call to
mod2pi() without affecting the results in the least.