Adjust test limit for 'long double' Perls.
I believe that the problem is that the reference data were generated
by a Fortran program using double precision rather than long double, so
we're really dealing with a difference in rounding/truncation error.
This really only affects Mercury, the problem child of the Solar System.

Thanks to Slaven Rezić (SREZIC) for testing this configuration and
exposing the issue.